+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 11/09/2025
+
+### Added
+
+- Added Image upload support for profile picture and project cover
+- Uses S3 for object storage
+- Added edit functions for Employment and Projects
+
 ## [0.1.3] - 11/08/2025
 
 ### Added